GADE PART 2 UPDATE
In this commit I have included environmental elements which fit perfectly for the theme of this runner game. I have also included a timer and added a functionality which will decrease the player's health every second by 10 which the player needs to collect health. Lastly, the death menu instructs what occurred and how the Player died, either by health or an obstacle.