GADE UPDATE
This is a commit to show all fixtures on the groundtile, pickups and obstacles spawning creating a more flexible way to handle the spawning of tiles.
The portal now functions properly and transports the player further enough in the game world. The physics of the game is properly fixed and placed.