GADE UPDATE
The obstacles and pickups do not initially overlap each other. The environment is perfectly implemented. An addition for another groundtile before collected any pickup had been instated to assist with any upcoming errors before the Player enters the actual game world ground.

The wooden spike finally moves on the x axis as the evil hands scale in size to show complexity. All material placed on the prefabs have been placed.

Currently working on the UI and sounds, including the events of the game