GADE UPDATE
The events system for obstacles and pickups finally work and officially functional. The environment is fixed with planets hanging in the sky giving a more game feel. The Ui is updated allowing for levels to interchange and for the controls panel to be placed. Left is the sounds to be added into the game to give a more enhanced experience